requirements.txt
flask==3.0.3
python-dotenv==1.0.1
requests==2.31.0
google-auth==2.29.0
google-api-python-client==2.125.0
